Add a simple add-on that straightens a bezier curve.
The immediately intended use case is for using curves to position
cloth ribbons before the start of simulation. The curve modifier
on its own distorts the shape of the ribbon, but it can be fixed
by adding a shape key for the curve that straightens it out, and
snapping the curve to it after the start of simulation with
Dynamic Base Mesh enabled in the cloth sim.